Move index routes into their own file
